renamed rule controller
