+++ src/routes/api/rules.js
@@ -1,5 +1,5 @@
 const router = require('express').Router();
-const RuleController = require('../../controllers/api/rulesController');
+const RuleController = require('../../controllers/api/ruleController');
 
 router.get('/', RuleController.index);
 router.get('/:index', RuleController.show);
